organization

+++ lib/derecho/rackspace/load_balancer.rb
@@ -1,43 +1,89 @@
 class Derecho
   module Rackspace
     class Load_Balancer
-      
+   
       def initialize
-        config = Derecho::Config.new.read
+        config = Config.new.read
         settings = config['accounts']['rackspace']
+        
+        puts $:
         
         @service = Fog::Rackspace::LoadBalancers.new({
           :rackspace_username    => settings['username'],
           :rackspace_api_key     => settings['api_key'],
           :rackspace_lb_endpoint => "https://#{settings['region']}.loadbalancers.api.rackspacecloud.com/v1.0/"
         })
-      end
-      
-      def create name, server_id, protocol = 'HTTP', port = 80, virtual_ip_type = 'PUBLIC'
-        server_ip = Derecho::Rackspace::Server.new.get(server_id).private_ip_address
         
-        lb = @service.load_balancers.create(
+        @lb = @service.load_balancers
+      end
+    
+      def create name, srv_id, protocol = 'HTTP', port = 80, virtual_ip_type = 'PUBLIC'
+        srv_ip = Rackspace::Server.new.get(srv_id).private_ip_address
+        
+        @lb.create(
           :name     => name,
           :protocol => protocol,
           :port     => port,
           :virtual_ips => [{:type => virtual_ip_type}],
-          :nodes => [{:address => server_ip, :condition => 'ENABLED', :port => port}]  
+          :nodes => [{:address => srv_ip, :condition => 'ENABLED', :port => port}]  
         )
       end
       
-      def delete lb_id
-        lb = get lb_id
-        lb.destroy
-        lb
+      def get lb_id
+        @lb.get lb_id
       end
       
       def all
-        @service.load_balancers.all
+        @lb.all
       end
       
-      def get lb_id
-        @service.load_balancers.get lb_id
+      def delete lb_id
+        lb = @lb.get lb_id
+        lb.destroy unless lb.nil?
+        lb
       end
+      
+      class << self
+        
+        def exists? lb_id
+          lb = self.new.get lb_id
+          !lb.nil?
+        end
+        
+        def srv_exists? srv_id
+          Rackspace::Server.exists? srv_id
+        end
+        
+      end
+      
+      def is_server_attached? lb_id, srv_id
+        if exists? lb_id and srv_exists? srv_id  
+          srv = Rackspace::Server.new.get server_id
+        
+          get_nodes(lb_id).any? do |node| 
+            node.address == srv.private_ip_address
+          end
+        end
+      end
+
+      def attach_server lb_id, server_id, port = nil, condition = 'ENABLED'
+        unless is_server_attached? lb_id, server_id
+          lb = get lb_id
+          srv = Rackspace::Server.new.get server_id
+          
+          lb.nodes.create(
+            :address => srv.private_ip_address,
+            :port => port || lb.port,
+            :condition => condition
+          ).save
+          
+          lb
+        else
+          raise "#{srv.name} is already attached to #{lb.name}"
+        end
+      end
+
+      # working with nodes
       
       def get_nodes lb_id
         lb = get lb_id
@@ -45,14 +91,8 @@ class Derecho
       end
       
       def get_node lb_id, node_id
-        nodes = get_nodes lb_id 
-        nodes.select { |node| node.id == node_id }
-        nodes.first
-      end
-      
-      def exists? lb_id
         lb = get lb_id
-        !lb.nil?
+        lb.nodes.get node_id
       end
       
       def node_exists? lb_id, node_id
@@ -60,43 +100,8 @@ class Derecho
         !node.nil?
       end
       
-      def server_exists? server_id
-        server = Derecho::Rackspace::Server.new.get server_id
-        !server.nil?
-      end
-      
-      def is_attached? lb_id, server_id
-        if server_exists? server_id
-          server = Derecho::Rackspace::Server.new.get server_id
-          get_nodes(lb_id).any? do |node| 
-            node.address == server.private_ip_address
-          end
-        else
-          false
-        end
-      end
-      
-      def attach lb_id, server_id, port = nil, condition = 'ENABLED'
-        # check that both lb and srv are valid
-        if exists? lb_id and server_exists? server_id
-          # ignore the request if it's already attached
-          unless is_attached? lb_id, server_id
-            lb = get lb_id
-            server = Derecho::Rackspace::Server.new.get server_id
-            
-            node = @service.nodes.new :service => @service, :load_balancer => lb
-            node.address = server.private_ip_address
-            node.port = port || lb.port
-            node.condition = condition
-            node.save
-            
-            lb
-          end
-        end
-      end
-      
-      def detach lb_id, node_id
-        if exists? lb_id
+      def detach_node lb_id, node_id
+        if is_node_attached? lb_id, node_id
           lb = get lb_id
           nodes = get_nodes lb_id
